Refactoring: move to dedicated directory and split into modules
