+++ test/integration/framework/requirements.txt
@@ -1,0 +1,2 @@
+PyYAML==5.4.1
+web3==5.23.1
