--- test/integration/project
@@ -1,13 +1,0 @@
-#!/bin/python3
-
-import os
-import sys
-
-basedir = os.path.abspath(os.path.join(os.path.normpath(os.path.join(__file__, *([os.path.pardir] * 3)))))
-venv_dir = os.path.join(basedir, "venv")
-test_integration_dir = os.path.join(basedir, "test/integration")
-
-sys.path = sys.path + [os.path.join(venv_dir, "lib/python3.6/site-packages"), os.path.join(test_integration_dir)]
-
-import make
-make.main(sys.argv[1:])
